+++ swak/cloud/aws/__init__.py
@@ -8,15 +8,16 @@ from .s3 import S3
 from .bucket import S3Bucket
 from .df2s3 import DataFrame2S3Parquet
 from .s32df import S3Parquet2DataFrame
+from .s32local import S3File2LocalFile
 
 __all__ = [
     'S3',
     'S3Bucket',
     'DataFrame2S3Parquet',
-    'S3Parquet2DataFrame'
+    'S3Parquet2DataFrame',
+    'S3File2LocalFile'
 ]
 
-# ToDo: Add download file to local!
 # ToDo: Add discover files!
 # ToDo: Add download directory to local?
 # ToDo: Add download directory to dataframe?